refactor: Add form submission functionality to AddTask component
Enable form submission in the AddTask component by implementing the handleSubmit function. This function sends a POST request to the server with the task data entered in the form. If the request is successful, the form is reset and the popup is closed. If there is an error, it is logged to the console.