refactor: Update CardNotes component to include delete functionality
Update the CardNotes component to include delete functionality. This involves adding a handleDelete function that sends a DELETE request to the server to delete the note with the specified ID. If the request is successful, the note is removed from the UI. If there is an error, it is logged to the console.