refactor: Update CardReminders component to include delete functionality
Update the CardReminders component to include delete functionality. This involves adding a handleDelete function that sends a DELETE request to the server to delete the reminder with the specified ID. If the request is successful, the reminder is removed from the UI. If there is an error, it is logged to the console.