refactor: Add onDelete functionality to CardTransactions component
Add the onDelete functionality to the CardTransactions component by including the onDelete prop and updating the handleDelete function. This allows the user to delete a transaction by sending a DELETE request to the server. If the request is successful, the transaction is removed from the UI and the onDelete callback is triggered. Any errors are logged to the console.