refactor: Add form submission functionality to Signup component
Enable form submission in the Signup component by implementing the handleSignup function. This function sends a POST request to the server with the email and password entered in the form. If the request is successful and the response contains a valid user ID, the user is navigated to the dashboard. If there is an error during the fetch or signup process, an error message is displayed. The password input field also includes validation to ensure it is at least 8 characters long.