refactor: Update RingingAlarm component to use Button component for close functionality
Refactor the RingingAlarm component to use the Button component for the close functionality. This improves the consistency and reusability of the code. The closePopup function is now triggered by the onClick event of the Button component. The Button component is styled with the "rounded-lg" class and has a variant of "secondary".